% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMR.R
\name{TEMR}
\alias{TEMR}
\title{TEMR}
\usage{
TEMR(betaXG, betaYG, sebetaXG, sebetaYG, rho, meth)
}
\arguments{
\item{betaXG}{nSNP*nRace matrix of G-X association (beta coefficient),
the last column is auxiliary population.}

\item{betaYG}{nSNP*nRace matrix of G-Y association (beta coefficient),
the last column is auxiliary population.}

\item{sebetaXG}{nSNP*nRace matrix of G-X association (standard error),
the last column is auxiliary population.}

\item{sebetaYG}{nSNP*nRace matrix of G-Y association (standard error),
the last column is auxiliary population.}

\item{rho}{nRace*nRace trans-ethnics genetic correlation matrix.}

\item{meth}{Optimization method to use, including "Nelder-Mead", "BFGS",
"CG", "L-BFGS-B", "SANN" and "Brent".}
}
\value{
A list object,
    resultall are the results of causal effect estimation in the
    target populations using MRTP;
    betaXY_IVW is the causal effect estimation of all populations using IVW;
    betaXY_Egger is the causal effect estimation of all populations using MR-Egger;
    betaXY_Final is the causal effect estimation of all populations
    removing the impact of pleiotropy.
}
\description{
Improving statistical power of causal effects in the one or
multiple target populations using a auxiliary population. Details refer to
TEMR: Trans-ethnic Mendelian Randomization Method using Large-scale GWAS
Summary Datasets by Lei Hou, Sijia Wu, Zhongshang Yuan, Hongkai Li, Fuzhong Xue
}
\examples{
betaXG <- NULL
sebetaXG <- NULL
sebetaYG <- NULL
for(i in 1:5){
  betaXG <- cbind(betaXG,
                  runif(100,0.05,0.2)
  )
  sebetaXG <- cbind(sebetaXG,
                    runif(100,0.01,0.1)
  )
  sebetaYG <- cbind(sebetaYG,
                    runif(100,0.01,0.1)
  )
}
betaYG <- betaXG*0.2 + rnorm(100)
race_name <- c('target1','target2','target3','target4','auxiliary')
colnames(betaXG) <- race_name
colnames(betaYG) <- race_name
colnames(sebetaXG) <- race_name
colnames(sebetaYG) <- race_name
rho <- NULL
meth <- 'CG'
TEMR(betaXG,betaYG,sebetaXG,sebetaYG,rho,meth)

}
